fix : 컨트롤러 상 불필요한 메서드 이동 및 간결화
