+++ src/main/java/gift/controller/MemberController.java
@@ -4,8 +4,6 @@ import gift.service.MemberService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.Base64;
-
 @RestController
 @RequestMapping("/api/members")
 public class MemberController {
@@ -17,28 +15,14 @@ public class MemberController {
 
     @PostMapping("/register")
     public ResponseEntity<String> signUp(@RequestHeader("Authorization") String str) {
-        var token = memberService.signUp(decodeToEmail(str), decodeToPassword(str));
+        var token = memberService.signUp(str);
 
         return ResponseEntity.ok(token);
     }
 
     @PostMapping("/login")
     public ResponseEntity<String> login(@RequestHeader("Authorization") String str) {
-        var token = memberService.login(decodeToEmail(str), decodeToPassword(str));
+        var token = memberService.login(str);
         return ResponseEntity.ok(token);
-    }
-
-    private String decodeToEmail(String str) {
-        String base64Credentials = str.substring("Basic ".length()).trim();
-        String credentials = new String(Base64.getDecoder().decode(base64Credentials));
-        String[] values = credentials.split(":", 2);
-        return values[0];
-    }
-
-    private String decodeToPassword(String str) {
-        String base64Credentials = str.substring("Basic ".length()).trim();
-        String credentials = new String(Base64.getDecoder().decode(base64Credentials));
-        String[] values = credentials.split(":", 2);
-        return values[1];
     }
 }
